refactor: 상품 도메인 별로 패키지 옮기기
