+++ src/main/java/kitchenpos/product/domain/ProductRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
+
+import kitchenpos.product.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
